aula 01

+++ aula_01/src/aula_01/DiferencaEntreValores.java
@@ -6,19 +6,19 @@ public class DiferencaEntreValores {
 	public static void main (String args[]) {
 		
 		Scanner entrada = new Scanner (System.in);
-		int numeroUm, numeroDois, numeroTres, numeroQuatro, diferenca;
+		float numeroUm, numeroDois, numeroTres, numeroQuatro, diferenca;
 		
 		System.out.println("Digite um número: ");
-		numeroUm = entrada.nextInt();
+		numeroUm = entrada.nextFloat();
 		
 		System.out.println("Digite outro número: ");
-		numeroDois = entrada.nextInt();
+		numeroDois = entrada.nextFloat();
 		
 		System.out.println("Digite outro número: ");
-		numeroTres = entrada.nextInt();
+		numeroTres = entrada.nextFloat();
 		
 		System.out.println("Digite outro número: ");
-		numeroQuatro = entrada.nextInt();
+		numeroQuatro = entrada.nextFloat();
 		
 		diferenca = (numeroUm * numeroDois) - (numeroTres * numeroQuatro);
